Summarize model fit

Provides a summary of:

Estimated kits distributed

Percentage of kits distributed that are reported

Estimated kits used

percentage of kits used that are reported

percentage of kits orders that are used

probability kit used if distributed

kit_summary_table(
  fit,
  ...,
  data = NULL,
  accuracy = 0.01,
  cri_range = 0.95,
  ndraws = NULL
)

Arguments

fit

stanfit object

...

variables to group by in estimate

data

data used for model fitting. Can also include p_use column which can be used to plot true values if derived from simulated data.

accuracy

A number to round to. Use (e.g.) 0.01 to show 2 decimal places of precision. If NULL, the default, uses a heuristic that should ensure breaks have the minimum number of digits needed to show the difference between adjacent values.

cri_range

The range of the credible interval e.g. 0.95

ndraws

Number of draws to use in estimate

Value

A tibble::tibble

Probability of kit use if distributed

Estimated as distributed

Proportion kits distributed that are reported

Estimated kits used

Proportion kits used that are reported

Proportion kits ordered that are used

See also

Examples

if (FALSE) {
  fit <- est_naloxone(d)
  kit_summary_table(fit,regions,data = d)
}