generate model data for testing purposes

Model generating process using random walk to match data generating model in Bayesian framework

model_random_walk_data(
  N_t = 24,
  region_coeffs = c(5, 0.5),
  c_region = c(-1, 2),
  sigma = 2,
  zeta = 0.5,
  mu0 = -1,
  Orders = NULL,
  reporting_freq = NULL
)

Arguments

N_t

number of time-points

region_coeffs

vector of coefficients for regions determining kit orders

c_region

logit probability of kit use per region

sigma

standard deviation of error in logit probability of kit use

zeta

standard deviation of random walk in logit space

mu0

initial condition of random walk in logit space

Orders

A 2D matrix of shape length(region_coeffs) by N_t

reporting_freq

The frequency that distribution data is provided. If NULL distribution frequency matches orders frequency

Value

A tibble

Orders

Kit orders per time and region

regions

Numeric index indicating region of orders and distributions

Reported_Used

Number of kits reported as used

Reported_Distributed

Number of kits reported as distributed

p_use

Probability that a kit was used

p_reported

Probability that a distributed kit was reported

times

Index for time

region_name

String index for the region

See also